Add an accept decorator
For route handlers that respond to the GET method, add an optional
accept decorator which raises a 406 response code if the accept
header in the request does not match with the media types provided
to the decorator. ACCEPTABLE_TYPES contains defaults.

This gets the basics in place but leaves room for later improvement
in how the body of the 406 response is formed. It could be more useful
if we desire.

In order to make this change in a tidy way that did not drive the
author crazy not route-handling methods have been moved out of
handler.py into decorators.py, errors.py and request_funcs.py depending
on what kind of function those methods provided.

Various gabbits have been updated to reflect the required accept header
and also correct the fact that I dumbly used content-type where I meant
accept.
