+++ enamel/api/handlers.py
@@ -11,87 +11,12 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-# NOTE(cdent): We need a refactoring to break this into smaller
-# modules soon.
-
 import os
 import re
 
 import flask
-import httpexceptor
-from oslo_utils import uuidutils
 
-from enamel.api import version
-
-
-REQUEST_ID_HEADER = 'Openstack-Request-ID'
-
-
-def handle_error(error):
-    """Trap an HTTPException and package it for display"""
-    # This current implementation is based on the httpexceptor model
-    # which provides a very simple way to raise an HTTP<some status>
-    # anywhere. This takes that exception, extracts meaning from it
-    # and puts it in the format of an OpenStack errors object.
-
-    status_code, title = error.status.split(' ', 1)
-    # httpexcepter body() is a one item list of bytestring
-    body_detail = error.body()[0].decode('utf-8')
-
-    response = flask.jsonify({'errors': [{
-        'status': int(status_code),
-        'request_id': flask.g.request_id,
-        'title': title,
-        'detail': body_detail,
-    }]})
-
-    response.status = error.status
-    for header, value in error.headers():
-        if header.lower() == 'content-type':
-            continue
-        response.headers[header] = value
-    return response
-
-
-def handle_404(error):
-    """Transform a Flask 404 into our error handling flow."""
-    return handle_error(httpexceptor.HTTP404(error.description))
-
-
-def set_request_id():
-    """A before_request function to set required-id."""
-    flask.g.request_id = uuidutils.generate_uuid()
-
-
-def send_request_id(response):
-    """An after_request_function to send request-id header."""
-    response.headers[REQUEST_ID_HEADER] = flask.g.request_id
-    return response
-
-
-def set_version():
-    """A before_request function to set microversion."""
-    try:
-        flask.g.request_version = version.extract_version(
-            flask.request.headers)
-    except ValueError as exc:
-        flask.g.request_version = version.parse_version_string(
-            version.min_version_string())
-        raise httpexceptor.HTTP406('unable to use provided version: %s' % exc)
-
-
-def send_version(response):
-    """An after_request function to send microversion headers."""
-    vary = response.headers.get('vary')
-    header = version.Version.HEADER
-    value = flask.g.get('request_version')
-    if value:
-        if vary:
-            response.headers['vary'] = '%s, %s' % (vary, header)
-        else:
-            response.headers['vary'] = header
-        response.headers[header] = '%s %s' % (version.SERVICE_TYPE, value)
-    return response
+from enamel.api.decorators import accept
 
 
 def create_link_object(urls):
@@ -112,6 +37,7 @@ def generate_resource_data(resources):
     return data
 
 
+@accept()
 def home():
     pat = re.compile("^\/[^\/]*?$")
 
@@ -123,6 +49,8 @@ def home():
     return flask.jsonify(resources=generate_resource_data(resources))
 
 
+# TODO(cdent): This should take a decorator like accept (above), but
+# for the content-type header.
 def server_boot():
     data = flask.request.get_json()
     # TODO(cdent): Call the task workflow here and return a link to the task
